backend/ai_agent/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
langchain==0.1.7
langchain-core==0.1.23
langchain-community==0.0.20
langchain-openai==0.0.6
langchain-google-genai==0.0.9
google-generativeai==0.3.2
openai==1.12.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
requests==2.31.0
regex==2023.12.25
tiktoken==0.6.0
marshmallow==3.20.2
typing_inspect==0.9.0
typing_extensions==4.9.0
PyPDF2==3.0.1
uvloop==0.19.0; sys_platform != 'win32'
watchfiles==0.21.0
colorama==0.4.6
tqdm==4.66.2
